v2.1/requirements.txt
cadquery-ocp>=7.7.0
PyQt6>=6.6.0
PyQt6-3D>=6.6.0
mujoco>=3.0.0
stable-baselines3>=2.2.0
trimesh>=4.0.0
numpy>=1.24.0
gymnasium>=0.29.0
